requirements.txt
celery
requests
redis
pyyaml
uuid